Create Robo Trumps
This code allows a user to choose a robot from a list and displays its stats. 
The data for each robot are stored in a text file named "cards.txt" and are read into the program at runtime.
The data is stored in a dictionary with the robot's name as the key and the data as a list of values. 
If the user inputs "random", the program will choose a random robot from the list. 
If the user inputs a valid robot name, the stats for that robot are displayed on the screen using turtle graphics'''